Move injected scripts to separate module
